Pas the result directory as a parameter for where to
put the result files.  Stripped out the system information overhead and
placed that in a separate file that is written only once per system.
handled the FAILURE notification by placing it in the record.  All new
records are placed at the end of the CSV file.  The csv files are
recorded by program (b2sum, dd, sha1sum, etc.)
so that the results file is not unweildy.