Big change with the /tmp when I discovered ...
that moving /tmp to tmpfs (RAM) can cause /tmp to be too
small to handle the number of copies of the dictionary.  This change
allows the user to override the percentage
of available space in /tmp that is consumed by the copies
of the dictionary.  It also provides the user with hints
on how to move /tmp to "/" if they are low on tmpfs space.